+++ src/shared/lib/updater/AutoUpdateBackgroundTask.js
@@ -2,11 +2,14 @@ import React from 'react';
 import { connect } from 'react-redux';
 import BackgroundTask from 'components/BackgroundTask';
 import { openConfirmDialog, showBackgroundTask } from 'actions/overlays';
+import { shell, remote } from 'electron';
 
 @connect(
   null,
-  openConfirmDialog,
-  showBackgroundTask
+  {
+    openConfirmDialog,
+    showBackgroundTask,
+  }
 )
 export default class AutoUpdateBackgroundTask extends React.Component {
   confirmInstall = () => {
@@ -22,14 +25,33 @@ export default class AutoUpdateBackgroundTask extends React.Component {
     });
   };
 
+  exitToGitHub = () => {
+    this.closeTask();
+    this.props.openConfirmDialog({
+      question: __('Close the wallet and go to GitHub?'),
+      labelYes: __('Close and go to Github'),
+      callbackYes: () => {
+        shell.openExternal(
+          'https://github.com/Nexusoft/NexusInterface/releases'
+        );
+        remote.app.quit();
+      },
+      labelNo: __('Install it later'),
+      callbackNo: () => {
+        this.props.showBackgroundTask(AutoUpdateBackgroundTask, this.props);
+      },
+    });
+  };
+
   render() {
+    const { gitHub } = this.props;
     return (
       <BackgroundTask
         type="success"
         assignClose={close => {
           this.closeTask = close;
         }}
-        onClick={this.confirmInstall}
+        onClick={gitHub ? this.exitToGitHub : this.confirmInstall}
       >
         {__('New wallet version %{version} - Ready to install!', {
           version: this.props.version,